refactor: Add InvalidCpfException and InvalidEmailException classes
