+++ src_temp/@Shared/InvalidCpfException.ts
@@ -1,0 +1,6 @@
+export default class InvalidCpfException extends Error {
+    constructor(message: string = 'Invalid Cpf') {
+        super(message)
+        this.name = 'InvalidCpfException'
+    }
+}
