+++ src_temp/@Shared/InvalidEmailException.ts
@@ -1,0 +1,6 @@
+export default class InvalidEmailException extends Error {
+    constructor(message: string = 'Invalid email') {
+        super(message)
+        this.name = 'InvalidEmailException'
+    }
+}
